renamed files and variable
